third_party/WebKit/LayoutTests/fast/dom/Range/deleteData-replaceData-count-overflow-expected.txt

Test deleteData() + replaceData() overflow handling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textNode = new Text('chopped off and then some'); textNode is non-null.
PASS textNode.deleteData(11, 0xfffffff7); textNode.data is "chopped off"
PASS range.endOffset is 11
PASS textNode = new Text('hello world'); textNode is non-null.
PASS textNode.replaceData(6, 0xfffffffe, 'bob'); textNode.data is "hello bob"
PASS range.endOffset is 6
PASS textNode = new Text('hello world'); textNode is non-null.
PASS textNode.replaceData(6, -1, 'bob'); textNode.data is "hello bob"
PASS textNode = new Text('hello world'); textNode is non-null.
PASS textNode.replaceData(6, -2, 'bob'); textNode.data is "hello bob"
PASS successfullyParsed is true

TEST COMPLETE
